Fixed bug when trying to delete players.

+++ ChessTourManager.DataAccess/Queries/Delete/DeleteQuery.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Windows;
 using ChessTourManager.DataAccess.Entities;
 using Microsoft.EntityFrameworkCore;
@@ -18,38 +19,55 @@ internal class DeleteQuery : IDeleteQueries
     {
         try
         {
+            if (CheckIfInGames(player))
+            {
+                MessageBox.Show("Нельзя удалить игрока, который участвует в игре!",
+                                "Ошибка при удалении игрока",
+                                MessageBoxButton.OK, MessageBoxImage.Error);
+                return DeleteResult.Failed;
+            }
+
             _context.Players.Remove(player);
             _context.SaveChanges();
+
             return DeleteResult.Success;
-        }
-        catch (InvalidOperationException)
-        {
-            MessageBox.Show("Нельзя удалить игрока, который участвует в игре!", 
-                            "Ошибка при удалении игрока", 
-                            MessageBoxButton.OK, MessageBoxImage.Error);
-            _context.Entry(player).State = EntityState.Unchanged;
-            return DeleteResult.Failed;
         }
         catch (Exception e)
         {
             MessageBox.Show(e.InnerException?.Message ?? e.Message, "Ошибка при удалении игрока",
                             MessageBoxButton.OK, MessageBoxImage.Error);
             _context.Entry(player).State = EntityState.Unchanged;
+
             return DeleteResult.Failed;
         }
+    }
+
+    private static bool CheckIfInGames(Player player)
+    {
+        bool isInWhiteGames = player.WhiteGamePlayers.Count != 0;
+        bool isInBlackGames = player.BlackGamePlayers.Count != 0;
+        return isInWhiteGames || isInBlackGames;
     }
 
     public DeleteResult TryDeleteTournament(Tournament tournament)
     {
         try
         {
+            // Drop cascade.
+            _context.Games
+                    .Where(g => g.OrganizerId  == tournament.OrganizerId
+                             && g.TournamentId == tournament.TournamentId)
+                    .ToList()
+                    .ForEach(g => _context.Games.Remove(g));
+
             _context.Tournaments.Remove(tournament);
             _context.SaveChanges();
             return DeleteResult.Success;
         }
-        catch (DbUpdateException)
+        catch (DbUpdateException e)
         {
-            MessageBox.Show("Нельзя удалить данный турнир!", "Ошибка при удалении турнира",
+            MessageBox.Show($"Нельзя удалить данный турнир!\n{e.InnerException?.Message}",
+                            "Ошибка при удалении турнира",
                             MessageBoxButton.OK, MessageBoxImage.Error);
             _context.Entry(tournament).State = EntityState.Unchanged;
             return DeleteResult.Failed;
